requirements.txt
requests==2.31.0
python-dotenv==1.0.1
tqdm==4.66.2
google-generativeai==0.3.2
anyio==4.8.0
openai==1.3.0
datetime 